+++ src/main/kotlin/lotto/ui/LottoGamePrinter.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.ui
 
 import lotto.domain.LottoRank
 import lotto.domain.LottoResults
